Clicking tab should clear search
